Исправить ошибки в преобразовании формата продолжительности трека и неточности в верстке экрана поиска и элемента списка треков

+++ app/src/main/java/com/example/playlistmaker/TrackViewHolder.kt
@@ -20,7 +20,7 @@ class TrackViewHolder(itemView: View): RecyclerView.ViewHolder(itemView) {
     fun bind(model: Track) {
         trackName.text = model.trackName
         artistName.text = model.artistName
-        trackTime.text = SimpleDateFormat("mm:ss", Locale.getDefault()).format(model.trackTimeMillis)
+        trackTime.text = SimpleDateFormat("mm:ss", Locale.getDefault()).format(model.trackTimeMillis.toLong())
 
         Glide.with(itemView)
             .load(model.artworkUrl100)
